ClawGuard-BNB/requirements.txt
# ClawGuard-BNB 核心依赖

# 核心依赖
requests>=2.31.0
pyyaml>=6.0.1
cryptography>=41.0.0

# WebSocket 实时行情
websockets>=12.0

# 代理支持
pysocks>=1.7.1           # SOCKS5 代理支持
requests[socks]>=2.31.0  # requests with SOCKS support

# HTTP API 服务器
flask>=2.3.0
flask-cors>=4.0.0

# 可选依赖（用于高级功能）
# aiohttp>=3.9.0        # 异步HTTP客户端（WebSocket代理支持）
# pandas>=2.1.0         # 数据分析
# numpy>=1.26.0         # 数值计算

# 开发依赖（安装时使用 pip install -r requirements-dev.txt）
# pytest>=7.4.0         # 测试框架
# pytest-cov>=4.1.0     # 测试覆盖率
# pytest-asyncio>=0.21.0 # 异步测试
# black>=23.0.0         # 代码格式化
# flake8>=6.1.0         # 代码检查
# mypy>=1.7.0           # 类型检查

# 注意：如果在中国大陆安装较慢，可以使用清华镜像源：
# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
